@@ -1,160 +1,149 @@
+import numba as nb
 import numpy as np
 
 
-class ReducedDenstiyMatrix:
-    """Reduced density matrix class."""
+@nb.jit(nopython=True)
+def RDM1(p: int, q: int, num_inactive_orbs: int, num_active_orbs: int, rdm1: np.ndarray) -> float:
+    r"""Get full space one-electron reduced density matrix element.
 
-    def __init__(
-        self,
-        num_inactive_orbs: int,
-        num_active_orbs: int,
-        num_virtual_orbs: int,
-        rdm1: np.ndarray,
-        rdm2: np.ndarray | None = None,
-    ) -> None:
-        """Initialize reduced density matrix class.
+    The only non-zero elements are:
 
-        Args:
-            num_inactive_orbs: Number of inactive orbitals in spatial basis.
-            num_active_orbs: Number of active orbitals in spatial basis.
-            num_virtual_orbs: Number of virtual orbitals in spatial basis.
-            rdm1: One-electron reduced density matrix in the active space.
-            rdm2: Two-electron reduced density matrix in the active space.
-        """
-        self.inactive_idx = []
-        self.active_idx = []
-        self.virtual_idx = []
-        for idx in range(num_inactive_orbs):
-            self.inactive_idx.append(idx)
-        for idx in range(num_inactive_orbs, num_inactive_orbs + num_active_orbs):
-            self.active_idx.append(idx)
-        for idx in range(
-            num_inactive_orbs + num_active_orbs,
-            num_inactive_orbs + num_active_orbs + num_virtual_orbs,
-        ):
-            self.virtual_idx.append(idx)
-        self.idx_shift = num_inactive_orbs
-        self.rdm1 = rdm1
-        self.rdm2 = rdm2
+    .. math::
+        \Gamma^{[1]}_{pq} = \left\{\begin{array}{ll}
+                            2\delta_{ij} & pq = ij\\
+                            \left<0\left|\hat{E}_{vw}\right|0\right> & pq = vw\\
+                            0 & \text{otherwise} \\
+                            \end{array} \right.
 
-    def RDM1(self, p: int, q: int) -> float:
-        r"""Get full space one-electron reduced density matrix element.
+    and the symmetry `\Gamma^{[1]}_{pq}=\Gamma^{[1]}_{qp}`:math:.
 
-        The only non-zero elements are:
+    Args:
+        p: Spatial orbital index.
+        q: Spatial orbital index.
+        num_inactive_orbs: Number of spatial inactive orbitals.
+        num_active_orbs: Number of spatial active orbitals.
+        rdm1: Active part of 1-RDM.
 
-        .. math::
-            \Gamma^{[1]}_{pq} = \left\{\begin{array}{ll}
-                                2\delta_{ij} & pq = ij\\
-                                \left<0\left|\hat{E}_{vw}\right|0\right> & pq = vw\\
-                                0 & \text{otherwise} \\
-                                \end{array} \right.
-
-        and the symmetry `\Gamma^{[1]}_{pq}=\Gamma^{[1]}_{qp}`:math:.
-
-        Args:
-            p: Spatial orbital index.
-            q: Spatial orbital index.
-
-        Returns:
-            One-electron reduced density matrix element.
-        """
-        if p in self.active_idx and q in self.active_idx:
-            return self.rdm1[p - self.idx_shift, q - self.idx_shift]
-        if p in self.inactive_idx and q in self.inactive_idx:
-            if p == q:
-                return 2
-            return 0
+    Returns:
+        One-electron reduced density matrix element.
+    """
+    virt_start = num_inactive_orbs + num_active_orbs
+    if p >= virt_start or q >= virt_start:
+        # Zero if any virtual index
         return 0
-
-    def RDM2(self, p: int, q: int, r: int, s: int) -> float:  # pylint: disable=R0911
-        r"""Get full space two-electron reduced density matrix element.
-
-        .. math::
-            \Gamma^{[2]}_{pqrs} = \left\{\begin{array}{ll}
-                                  4\delta_{ij}\delta_{kl} - 2\delta_{jk}\delta_{il} & pqrs = ijkl\\
-                                  2\delta_{ij} \Gamma^{[1]}_{vw} & pqrs = vwij\\
-                                  - \delta_{ij}\Gamma^{[1]}_{vw} & pqrs = ivwj\\
-                                  \left<0\left|\hat{e}_{vwxy}\right|0\right> & pqrs = vwxy\\
-                                  0 & \text{otherwise} \\
-                                  \end{array} \right.
-
-        and the symmetry `\Gamma^{[2]}_{pqrs}=\Gamma^{[2]}_{rspq}=\Gamma^{[2]}_{qpsr}=\Gamma^{[2]}_{srqp}`:math:.
-
-        Args:
-            p: Spatial orbital index.
-            q: Spatial orbital index.
-            r: Spatial orbital index.
-            s: Spatial orbital index.
-
-        Returns:
-            Two-electron reduced density matrix element.
-        """
-        if self.rdm2 is None:
-            raise ValueError("RDM2 is not given.")
-        if p in self.active_idx and q in self.active_idx and r in self.active_idx and s in self.active_idx:
-            return self.rdm2[
-                p - self.idx_shift,
-                q - self.idx_shift,
-                r - self.idx_shift,
-                s - self.idx_shift,
-            ]
-        if (
-            p in self.inactive_idx
-            and q in self.active_idx
-            and r in self.active_idx
-            and s in self.inactive_idx
-        ):
-            if p == s:
-                return -self.rdm1[q - self.idx_shift, r - self.idx_shift]
-            return 0
-        if (
-            p in self.active_idx
-            and q in self.inactive_idx
-            and r in self.inactive_idx
-            and s in self.active_idx
-        ):
-            if q == r:
-                return -self.rdm1[p - self.idx_shift, s - self.idx_shift]
-            return 0
-        if (
-            p in self.active_idx
-            and q in self.active_idx
-            and r in self.inactive_idx
-            and s in self.inactive_idx
-        ):
-            if r == s:
-                return 2 * self.rdm1[p - self.idx_shift, q - self.idx_shift]
-            return 0
-        if (
-            p in self.inactive_idx
-            and q in self.inactive_idx
-            and r in self.active_idx
-            and s in self.active_idx
-        ):
-            if p == q:
-                return 2 * self.rdm1[r - self.idx_shift, s - self.idx_shift]
-            return 0
-        if (
-            p in self.inactive_idx
-            and q in self.inactive_idx
-            and r in self.inactive_idx
-            and s in self.inactive_idx
-        ):
-            val = 0
-            if p == q and r == s:
-                val += 4
-            if q == r and p == s:
-                val -= 2
-            return val
+    elif p >= num_inactive_orbs and q >= num_inactive_orbs:
+        # All active index
+        return rdm1[p - num_inactive_orbs, q - num_inactive_orbs]
+    elif p < num_inactive_orbs and q < num_inactive_orbs:
+        # All inactive indx
+        if p == q:
+            return 2
         return 0
+    # One inactive and one active index
+    return 0
 
 
+@nb.jit(nopython=True)
+def RDM2(
+    p: int,
+    q: int,
+    r: int,
+    s: int,
+    num_inactive_orbs: int,
+    num_active_orbs: int,
+    rdm1: np.ndarray,
+    rdm2: np.ndarray,
+) -> float:
+    r"""Get full space two-electron reduced density matrix element.
+
+    .. math::
+        \Gamma^{[2]}_{pqrs} = \left\{\begin{array}{ll}
+                              4\delta_{ij}\delta_{kl} - 2\delta_{jk}\delta_{il} & pqrs = ijkl\\
+                              2\delta_{ij} \Gamma^{[1]}_{vw} & pqrs = vwij\\
+                              - \delta_{ij}\Gamma^{[1]}_{vw} & pqrs = ivwj\\
+                              \left<0\left|\hat{e}_{vwxy}\right|0\right> & pqrs = vwxy\\
+                              0 & \text{otherwise} \\
+                              \end{array} \right.
+
+    and the symmetry `\Gamma^{[2]}_{pqrs}=\Gamma^{[2]}_{rspq}=\Gamma^{[2]}_{qpsr}=\Gamma^{[2]}_{srqp}`:math:.
+
+    Args:
+        p: Spatial orbital index.
+        q: Spatial orbital index.
+        r: Spatial orbital index.
+        s: Spatial orbital index.
+        num_inactive_orbs: Number of spatial inactive orbitals.
+        num_active_orbs: Number of spatial active orbitals.
+        rdm1: Active part of 1-RDM.
+        rdm2: Active part of 2-RDM.
+
+    Returns:
+        Two-electron reduced density matrix element.
+    """
+    virt_start = num_inactive_orbs + num_active_orbs
+    if p >= virt_start or q >= virt_start or r >= virt_start or s >= virt_start:
+        # Zero if any virtual index
+        return 0
+    elif (
+        p >= num_inactive_orbs
+        and q >= num_inactive_orbs
+        and r >= num_inactive_orbs
+        and s >= num_inactive_orbs
+    ):
+        return rdm2[
+            p - num_inactive_orbs,
+            q - num_inactive_orbs,
+            r - num_inactive_orbs,
+            s - num_inactive_orbs,
+        ]
+    elif (
+        p < num_inactive_orbs and q >= num_inactive_orbs and r >= num_inactive_orbs and s < num_inactive_orbs
+    ):
+        # iuvj type index
+        if p == s:
+            return -rdm1[q - num_inactive_orbs, r - num_inactive_orbs]
+        return 0
+    elif (
+        p >= num_inactive_orbs and q < num_inactive_orbs and r < num_inactive_orbs and s >= num_inactive_orbs
+    ):
+        # uijv type index
+        if q == r:
+            return -rdm1[p - num_inactive_orbs, s - num_inactive_orbs]
+        return 0
+    elif (
+        p >= num_inactive_orbs and q >= num_inactive_orbs and r < num_inactive_orbs and s < num_inactive_orbs
+    ):
+        # uvij type index
+        if r == s:
+            return 2 * rdm1[p - num_inactive_orbs, q - num_inactive_orbs]
+        return 0
+    elif (
+        p < num_inactive_orbs and q < num_inactive_orbs and r >= num_inactive_orbs and s >= num_inactive_orbs
+    ):
+        # ijuv type index
+        if p == q:
+            return 2 * rdm1[r - num_inactive_orbs, s - num_inactive_orbs]
+        return 0
+    elif p < num_inactive_orbs and q < num_inactive_orbs and r < num_inactive_orbs and s < num_inactive_orbs:
+        # All inactive index
+        val = 0
+        if p == q and r == s:
+            val += 4
+        if q == r and p == s:
+            val -= 2
+        return val
+    # Everything else
+    return 0
+
+
+@nb.jit(nopython=True)
 def get_electronic_energy(
-    rdms: ReducedDenstiyMatrix,
     h_int: np.ndarray,
     g_int: np.ndarray,
     num_inactive_orbs: int,
     num_active_orbs: int,
+    rdm1: np.ndarray,
+    rdm2: np.ndarray,
 ) -> float:
     r"""Calculate electronic energy.
 
@@ -166,6 +155,8 @@ def get_electronic_energy(
         g_int: Two-electron integrals in MO.
         num_inactive_orbs: Number of inactive orbitals.
         num_active_orbs: Number of active orbitals.
+        rdm1: Active part of 1-RDM.
+        rdm2: Active part of 2-RDM.
 
     Returns:
         Electronic energy.
@@ -173,22 +164,29 @@ def get_electronic_energy(
     energy = 0
     for p in range(num_inactive_orbs + num_active_orbs):
         for q in range(num_inactive_orbs + num_active_orbs):
-            energy += h_int[p, q] * rdms.RDM1(p, q)
+            energy += h_int[p, q] * RDM1(p, q, num_inactive_orbs, num_active_orbs, rdm1)
     for p in range(num_inactive_orbs + num_active_orbs):
         for q in range(num_inactive_orbs + num_active_orbs):
             for r in range(num_inactive_orbs + num_active_orbs):
                 for s in range(num_inactive_orbs + num_active_orbs):
-                    energy += 1 / 2 * g_int[p, q, r, s] * rdms.RDM2(p, q, r, s)
+                    energy += (
+                        1
+                        / 2
+                        * g_int[p, q, r, s]
+                        * RDM2(p, q, r, s, num_inactive_orbs, num_active_orbs, rdm1, rdm2)
+                    )
     return energy
 
 
+@nb.jit(nopython=True)
 def get_orbital_gradient(
-    rdms: ReducedDenstiyMatrix,
     h_int: np.ndarray,
     g_int: np.ndarray,
-    kappa_idx: list[list[int]],
+    kappa_idx: list[tuple[int, int]],
     num_inactive_orbs: int,
     num_active_orbs: int,
+    rdm1: np.ndarray,
+    rdm2: np.ndarray,
 ) -> np.ndarray:
     r"""Calculate the orbital gradient.
 
@@ -196,12 +194,13 @@ def get_orbital_gradient(
         g_{pq}^{\hat{\kappa}} = \left<0\left|\left[\hat{\kappa}_{pq},\hat{H}\right]\right|0\right>
 
     Args:
-       rdms: Reduced density matrix class.
-       h_int: One-electron integrals in MO in Hamiltonian.
-       g_int: Two-electron integrals in MO in Hamiltonian.
-       kappa_idx: Orbital parameter indicies in spatial basis.
-       num_inactive_orbs: Number of inactive orbitals in spatial basis.
-       num_active_orbs: Number of active orbitals in spatial basis.
+        h_int: One-electron integrals in MO in Hamiltonian.
+        g_int: Two-electron integrals in MO in Hamiltonian.
+        kappa_idx: Orbital parameter indices in spatial basis.
+        num_inactive_orbs: Number of inactive orbitals in spatial basis.
+        num_active_orbs: Number of active orbitals in spatial basis.
+        rdm1: Active part of 1-RDM.
+        rdm2: Active part of 2-RDM.
 
     Returns:
         Orbital gradient.
@@ -210,26 +209,36 @@ def get_orbital_gradient(
     for idx, (m, n) in enumerate(kappa_idx):
         # 1e contribution
         for p in range(num_inactive_orbs + num_active_orbs):
-            gradient[idx] += 2 * h_int[n, p] * rdms.RDM1(m, p)
-            gradient[idx] -= 2 * h_int[p, m] * rdms.RDM1(p, n)
+            gradient[idx] += 2 * h_int[n, p] * RDM1(m, p, num_inactive_orbs, num_active_orbs, rdm1)
+            gradient[idx] -= 2 * h_int[p, m] * RDM1(p, n, num_inactive_orbs, num_active_orbs, rdm1)
         # 2e contribution
         for p in range(num_inactive_orbs + num_active_orbs):
             for q in range(num_inactive_orbs + num_active_orbs):
                 for r in range(num_inactive_orbs + num_active_orbs):
-                    gradient[idx] += g_int[n, p, q, r] * rdms.RDM2(m, p, q, r)
-                    gradient[idx] -= g_int[p, m, q, r] * rdms.RDM2(p, n, q, r)
-                    gradient[idx] -= g_int[m, p, q, r] * rdms.RDM2(n, p, q, r)
-                    gradient[idx] += g_int[p, n, q, r] * rdms.RDM2(p, m, q, r)
+                    gradient[idx] += g_int[n, p, q, r] * RDM2(
+                        m, p, q, r, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
+                    gradient[idx] -= g_int[p, m, q, r] * RDM2(
+                        p, n, q, r, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
+                    gradient[idx] -= g_int[m, p, q, r] * RDM2(
+                        n, p, q, r, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
+                    gradient[idx] += g_int[p, n, q, r] * RDM2(
+                        p, m, q, r, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
     return gradient
 
 
+@nb.jit(nopython=True)
 def get_orbital_gradient_response(
-    rdms: ReducedDenstiyMatrix,
     h_int: np.ndarray,
     g_int: np.ndarray,
-    kappa_idx: list[list[int]],
+    kappa_idx: list[tuple[int, int]],
     num_inactive_orbs: int,
     num_active_orbs: int,
+    rdm1: np.ndarray,
+    rdm2: np.ndarray,
 ) -> np.ndarray:
     r"""Calculate the response orbital parameter gradient.
 
@@ -237,12 +246,13 @@ def get_orbital_gradient_response(
         g_{pq}^{\hat{q}} = \left<0\left|\left[\hat{q}_{pq},\hat{H}\right]\right|0\right>
 
     Args:
-       rdms: Reduced density matrix class.
-       h_int: One-electron integrals in MO in Hamiltonian.
-       g_int: Two-electron integrals in MO in Hamiltonian.
-       kappa_idx: Orbital parameter indicies in spatial basis.
-       num_inactive_orbs: Number of inactive orbitals in spatial basis.
-       num_active_orbs: Number of active orbitals in spatial basis.
+        h_int: One-electron integrals in MO in Hamiltonian.
+        g_int: Two-electron integrals in MO in Hamiltonian.
+        kappa_idx: Orbital parameter indices in spatial basis.
+        num_inactive_orbs: Number of inactive orbitals in spatial basis.
+        num_active_orbs: Number of active orbitals in spatial basis.
+        rdm1: Active part of 1-RDM.
+        rdm2: Active part of 2-RDM.
 
     Returns:
         Orbital response parameter gradient.
@@ -251,42 +261,90 @@ def get_orbital_gradient_response(
     for idx, (m, n) in enumerate(kappa_idx):
         # 1e contribution
         for p in range(num_inactive_orbs + num_active_orbs):
-            gradient[idx] += h_int[n, p] * rdms.RDM1(m, p)
-            gradient[idx] -= h_int[p, m] * rdms.RDM1(p, n)
+            gradient[idx] += h_int[n, p] * RDM1(m, p, num_inactive_orbs, num_active_orbs, rdm1)
+            gradient[idx] -= h_int[p, m] * RDM1(p, n, num_inactive_orbs, num_active_orbs, rdm1)
         # 2e contribution
         for p in range(num_inactive_orbs + num_active_orbs):
             for q in range(num_inactive_orbs + num_active_orbs):
                 for r in range(num_inactive_orbs + num_active_orbs):
-                    gradient[idx] += 1 / 2 * g_int[n, p, q, r] * rdms.RDM2(m, p, q, r)
-                    gradient[idx] -= 1 / 2 * g_int[p, m, q, r] * rdms.RDM2(p, n, q, r)
-                    gradient[idx] -= 1 / 2 * g_int[m, p, q, r] * rdms.RDM2(n, p, q, r)
-                    gradient[idx] += 1 / 2 * g_int[p, n, q, r] * rdms.RDM2(p, m, q, r)
+                    gradient[idx] += (
+                        1
+                        / 2
+                        * g_int[n, p, q, r]
+                        * RDM2(m, p, q, r, num_inactive_orbs, num_active_orbs, rdm1, rdm2)
+                    )
+                    gradient[idx] -= (
+                        1
+                        / 2
+                        * g_int[p, m, q, r]
+                        * RDM2(p, n, q, r, num_inactive_orbs, num_active_orbs, rdm1, rdm2)
+                    )
+                    gradient[idx] -= (
+                        1
+                        / 2
+                        * g_int[m, p, q, r]
+                        * RDM2(n, p, q, r, num_inactive_orbs, num_active_orbs, rdm1, rdm2)
+                    )
+                    gradient[idx] += (
+                        1
+                        / 2
+                        * g_int[p, n, q, r]
+                        * RDM2(p, m, q, r, num_inactive_orbs, num_active_orbs, rdm1, rdm2)
+                    )
     shift = len(kappa_idx)
     for idx, (n, m) in enumerate(kappa_idx):
         # 1e contribution
         for p in range(num_inactive_orbs + num_active_orbs):
-            gradient[idx + shift] += h_int[n, p] * rdms.RDM1(m, p)
-            gradient[idx + shift] -= h_int[p, m] * rdms.RDM1(p, n)
+            gradient[idx + shift] += h_int[n, p] * RDM1(m, p, num_inactive_orbs, num_active_orbs, rdm1)
+            gradient[idx + shift] -= h_int[p, m] * RDM1(p, n, num_inactive_orbs, num_active_orbs, rdm1)
         # 2e contribution
         for p in range(num_inactive_orbs + num_active_orbs):
             for q in range(num_inactive_orbs + num_active_orbs):
                 for r in range(num_inactive_orbs + num_active_orbs):
-                    gradient[idx + shift] += 1 / 2 * g_int[n, p, q, r] * rdms.RDM2(m, p, q, r)
-                    gradient[idx + shift] -= 1 / 2 * g_int[p, m, q, r] * rdms.RDM2(p, n, q, r)
-                    gradient[idx + shift] -= 1 / 2 * g_int[m, p, q, r] * rdms.RDM2(n, p, q, r)
-                    gradient[idx + shift] += 1 / 2 * g_int[p, n, q, r] * rdms.RDM2(p, m, q, r)
+                    gradient[idx + shift] += (
+                        1
+                        / 2
+                        * g_int[n, p, q, r]
+                        * RDM2(m, p, q, r, num_inactive_orbs, num_active_orbs, rdm1, rdm2)
+                    )
+                    gradient[idx + shift] -= (
+                        1
+                        / 2
+                        * g_int[p, m, q, r]
+                        * RDM2(p, n, q, r, num_inactive_orbs, num_active_orbs, rdm1, rdm2)
+                    )
+                    gradient[idx + shift] -= (
+                        1
+                        / 2
+                        * g_int[m, p, q, r]
+                        * RDM2(n, p, q, r, num_inactive_orbs, num_active_orbs, rdm1, rdm2)
+                    )
+                    gradient[idx + shift] += (
+                        1
+                        / 2
+                        * g_int[p, n, q, r]
+                        * RDM2(p, m, q, r, num_inactive_orbs, num_active_orbs, rdm1, rdm2)
+                    )
     return 2 ** (-1 / 2) * gradient
 
 
-def get_orbital_response_metric_sigma(rdms: ReducedDenstiyMatrix, kappa_idx: list[list[int]]) -> np.ndarray:
+@nb.jit(nopython=True)
+def get_orbital_response_metric_sigma(
+    kappa_idx: list[tuple[int, int]],
+    num_inactive_orbs: int,
+    num_active_orbs: int,
+    rdm1: np.ndarray,
+) -> np.ndarray:
     r"""Calculate the Sigma matrix orbital-orbital block.
 
     .. math::
         \Sigma_{pq,pq}^{\hat{q},\hat{q}} = \left<0\left|\left[\hat{q}_{pq}^\dagger,\hat{q}_{pq}\right]\right|0\right>
 
     Args:
-       rdms: Reduced density matrix class.
-       kappa_idx: Orbital parameter indicies in spatial basis.
+        kappa_idx: Orbital parameter indices in spatial basis.
+        num_inactive_orbs: Number of inactive orbitals in spatial basis.
+        num_active_orbs: Number of active orbitals in spatial basis.
+        rdm1: Active part of 1-RDM.
 
     Returns:
         Sigma matrix orbital-orbital block.
@@ -295,15 +353,18 @@ def get_orbital_response_metric_sigma(rdms: ReducedDenstiyMatrix, kappa_idx: lis
     for idx1, (n, m) in enumerate(kappa_idx):
         for idx2, (p, q) in enumerate(kappa_idx):
             if p == n:
-                sigma[idx1, idx2] += rdms.RDM1(m, q)
+                sigma[idx1, idx2] += RDM1(m, q, num_inactive_orbs, num_active_orbs, rdm1)
             if m == q:
-                sigma[idx1, idx2] -= rdms.RDM1(p, n)
+                sigma[idx1, idx2] -= RDM1(p, n, num_inactive_orbs, num_active_orbs, rdm1)
     return -1 / 2 * sigma
 
 
+@nb.jit(nopython=True)
 def get_orbital_response_vector_norm(
-    rdms: ReducedDenstiyMatrix,
     kappa_idx: list[list[int]],
+    num_inactive_orbs: int,
+    num_active_orbs: int,
+    rdm1: np.ndarray,
     response_vectors: np.ndarray,
     state_number: int,
     number_excitations: int,
@@ -314,11 +375,13 @@ def get_orbital_response_vector_norm(
         N^{\hat{q}} = \sum_k\left<0\left|\left[\hat{O}_{k},\hat{O}_{k}^\dagger\right]\right|0\right>
 
     Args:
-       rdms: Reduced density matrix class.
-       kappa_idx: Orbital parameter indicies in spatial basis.
-       response_vectors: Response vectors.
-       state_number: State number counting from zero.
-       number_excitations: Total number of excitations.
+        kappa_idx: Orbital parameter indices in spatial basis.
+        num_inactive_orbs: Number of inactive orbitals in spatial basis.
+        num_active_orbs: Number of active orbitals in spatial basis.
+        rdm1: Active part of 1-RDM.
+        response_vectors: Response vectors.
+        state_number: State number counting from zero.
+        number_excitations: Total number of excitations.
 
     Returns:
         Orbital part of excited state norm.
@@ -328,33 +391,38 @@ def get_orbital_response_vector_norm(
         for j, (t, u) in enumerate(kappa_idx):
             if n == u:
                 norm += (
-                    response_vectors[i, state_number] * response_vectors[j, state_number] * rdms.RDM1(m, t)
+                    response_vectors[i, state_number]
+                    * response_vectors[j, state_number]
+                    * RDM1(m, t, num_inactive_orbs, num_active_orbs, rdm1)
                 )
             if m == t:
                 norm -= (
-                    response_vectors[i, state_number] * response_vectors[j, state_number] * rdms.RDM1(n, u)
+                    response_vectors[i, state_number]
+                    * response_vectors[j, state_number]
+                    * RDM1(n, u, num_inactive_orbs, num_active_orbs, rdm1)
                 )
             if m == t:
                 norm += (
                     response_vectors[i + number_excitations, state_number]
                     * response_vectors[j + number_excitations, state_number]
-                    * rdms.RDM1(n, u)
+                    * RDM1(n, u, num_inactive_orbs, num_active_orbs, rdm1)
                 )
             if n == u:
                 norm -= (
                     response_vectors[i + number_excitations, state_number]
                     * response_vectors[j + number_excitations, state_number]
-                    * rdms.RDM1(m, t)
+                    * RDM1(m, t, num_inactive_orbs, num_active_orbs, rdm1)
                 )
     return 1 / 2 * norm
 
 
+@nb.jit(nopython=True)
 def get_orbital_response_property_gradient(
-    rdms: ReducedDenstiyMatrix,
     x_mo: np.ndarray,
-    kappa_idx: list[list[int]],
+    kappa_idx: list[tuple[int, int]],
     num_inactive_orbs: int,
     num_active_orbs: int,
+    rdm1: np.ndarray,
     response_vectors: np.ndarray,
     state_number: int,
     number_excitations: int,
@@ -365,14 +433,14 @@ def get_orbital_response_property_gradient(
         P^{\hat{q}} = \sum_k\left<0\left|\left[\hat{O}_{k},\hat{X}\right]\right|0\right>
 
     Args:
-       rdms: Reduced density matrix class.
-       x_mo: Property integral in MO basis.
-       kappa_idx: Orbital parameter indicies in spatial basis.
-       num_inactive_orbs: Number of inactive orbitals in spatial basis.
-       num_active_orbs: Number of active orbitals in spatial basis.
-       response_vectors: Response vectors.
-       state_number: State number counting from zero.
-       number_excitations: Total number of excitations.
+        x_mo: Property integral in MO basis.
+        kappa_idx: Orbital parameter indices in spatial basis.
+        num_inactive_orbs: Number of inactive orbitals in spatial basis.
+        num_active_orbs: Number of active orbitals in spatial basis.
+        rdm1: Active part of 1-RDM.
+        response_vectors: Response vectors.
+        state_number: State number counting from zero.
+        number_excitations: Total number of excitations.
 
     Returns:
         Orbital part of property gradient.
@@ -383,24 +451,26 @@ def get_orbital_response_property_gradient(
             prop_grad += (
                 (response_vectors[i + number_excitations, state_number] - response_vectors[i, state_number])
                 * x_mo[n, p]
-                * rdms.RDM1(m, p)
+                * RDM1(m, p, num_inactive_orbs, num_active_orbs, rdm1)
             )
             prop_grad += (
                 (response_vectors[i, state_number] - response_vectors[i + number_excitations, state_number])
                 * x_mo[m, p]
-                * rdms.RDM1(n, p)
+                * RDM1(n, p, num_inactive_orbs, num_active_orbs, rdm1)
             )
     return 2 ** (-1 / 2) * prop_grad
 
 
+@nb.jit(nopython=True)
 def get_orbital_response_hessian_block(
-    rdms: ReducedDenstiyMatrix,
     h: np.ndarray,
     g: np.ndarray,
-    kappa_idx1: list[list[int]],
-    kappa_idx2: list[list[int]],
+    kappa_idx1: list[tuple[int, int]],
+    kappa_idx2: list[tuple[int, int]],
     num_inactive_orbs: int,
     num_active_orbs: int,
+    rdm1: np.ndarray,
+    rdm2: np.ndarray,
 ) -> np.ndarray:
     r"""Calculate Hessian-like orbital-orbital block.
 
@@ -408,11 +478,14 @@ def get_orbital_response_hessian_block(
         H^{\hat{q},\hat{q}}_{tu,mn} = \left<0\left|\left[\hat{q}_{tu},\left[\hat{H},\hat{q}_{mn}\right]\right]\right|0\right>
 
     Args:
-       rdms: Reduced density matrix class.
-       kappa_idx1: Orbital parameter indicies in spatial basis.
-       kappa_idx1: Orbital parameter indicies in spatial basis.
-       num_inactive_orbs: Number of inactive orbitals in spatial basis.
-       num_active_orbs: Number of active orbitals in spatial basis.
+        h: Hamiltonian one-electron integrals in MO basis.
+        g: Hamiltonian two-electron integrals in MO basis.
+        kappa_idx1: Orbital parameter indices in spatial basis.
+        kappa_idx2: Orbital parameter indices in spatial basis.
+        num_inactive_orbs: Number of inactive orbitals in spatial basis.
+        num_active_orbs: Number of active orbitals in spatial basis.
+        rdm1: Active part of 1-RDM.
+        rdm2: Active part of 2-RDM.
 
     Returns:
         Hessian-like orbital-orbital block.
@@ -422,37 +495,69 @@ def get_orbital_response_hessian_block(
     for idx1, (t, u) in enumerate(kappa_idx1):
         for idx2, (m, n) in enumerate(kappa_idx2):
             # 1e contribution
-            A1e[idx1, idx2] += h[n, t] * rdms.RDM1(m, u)
-            A1e[idx1, idx2] += h[u, m] * rdms.RDM1(t, n)
+            A1e[idx1, idx2] += h[n, t] * RDM1(m, u, num_inactive_orbs, num_active_orbs, rdm1)
+            A1e[idx1, idx2] += h[u, m] * RDM1(t, n, num_inactive_orbs, num_active_orbs, rdm1)
             for p in range(num_inactive_orbs + num_active_orbs):
                 if m == u:
-                    A1e[idx1, idx2] -= h[n, p] * rdms.RDM1(t, p)
+                    A1e[idx1, idx2] -= h[n, p] * RDM1(t, p, num_inactive_orbs, num_active_orbs, rdm1)
                 if t == n:
-                    A1e[idx1, idx2] -= h[p, m] * rdms.RDM1(p, u)
+                    A1e[idx1, idx2] -= h[p, m] * RDM1(p, u, num_inactive_orbs, num_active_orbs, rdm1)
             # 2e contribution
             for p in range(num_inactive_orbs + num_active_orbs):
                 for q in range(num_inactive_orbs + num_active_orbs):
-                    A2e[idx1, idx2] += g[n, t, p, q] * rdms.RDM2(m, u, p, q)
-                    A2e[idx1, idx2] -= g[n, p, u, q] * rdms.RDM2(m, p, t, q)
-                    A2e[idx1, idx2] += g[n, p, q, t] * rdms.RDM2(m, p, q, u)
-                    A2e[idx1, idx2] += g[u, m, p, q] * rdms.RDM2(t, n, p, q)
-                    A2e[idx1, idx2] += g[p, m, u, q] * rdms.RDM2(p, n, t, q)
-                    A2e[idx1, idx2] -= g[p, m, q, t] * rdms.RDM2(p, n, q, u)
-                    A2e[idx1, idx2] -= g[u, p, n, q] * rdms.RDM2(t, p, m, q)
-                    A2e[idx1, idx2] += g[p, t, n, q] * rdms.RDM2(p, u, m, q)
-                    A2e[idx1, idx2] += g[p, q, n, t] * rdms.RDM2(p, q, m, u)
-                    A2e[idx1, idx2] += g[u, p, q, m] * rdms.RDM2(t, p, q, n)
-                    A2e[idx1, idx2] -= g[p, t, q, m] * rdms.RDM2(p, u, q, n)
-                    A2e[idx1, idx2] += g[p, q, u, m] * rdms.RDM2(p, q, t, n)
+                    A2e[idx1, idx2] += g[n, t, p, q] * RDM2(
+                        m, u, p, q, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
+                    A2e[idx1, idx2] -= g[n, p, u, q] * RDM2(
+                        m, p, t, q, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
+                    A2e[idx1, idx2] += g[n, p, q, t] * RDM2(
+                        m, p, q, u, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
+                    A2e[idx1, idx2] += g[u, m, p, q] * RDM2(
+                        t, n, p, q, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
+                    A2e[idx1, idx2] += g[p, m, u, q] * RDM2(
+                        p, n, t, q, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
+                    A2e[idx1, idx2] -= g[p, m, q, t] * RDM2(
+                        p, n, q, u, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
+                    A2e[idx1, idx2] -= g[u, p, n, q] * RDM2(
+                        t, p, m, q, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
+                    A2e[idx1, idx2] += g[p, t, n, q] * RDM2(
+                        p, u, m, q, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
+                    A2e[idx1, idx2] += g[p, q, n, t] * RDM2(
+                        p, q, m, u, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
+                    A2e[idx1, idx2] += g[u, p, q, m] * RDM2(
+                        t, p, q, n, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
+                    A2e[idx1, idx2] -= g[p, t, q, m] * RDM2(
+                        p, u, q, n, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
+                    A2e[idx1, idx2] += g[p, q, u, m] * RDM2(
+                        p, q, t, n, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                    )
             for p in range(num_inactive_orbs + num_active_orbs):
                 for q in range(num_inactive_orbs + num_active_orbs):
                     for r in range(num_inactive_orbs + num_active_orbs):
                         if m == u:
-                            A2e[idx1, idx2] -= g[n, p, q, r] * rdms.RDM2(t, p, q, r)
+                            A2e[idx1, idx2] -= g[n, p, q, r] * RDM2(
+                                t, p, q, r, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                            )
                         if t == n:
-                            A2e[idx1, idx2] -= g[p, m, q, r] * rdms.RDM2(p, u, q, r)
+                            A2e[idx1, idx2] -= g[p, m, q, r] * RDM2(
+                                p, u, q, r, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                            )
                         if m == u:
-                            A2e[idx1, idx2] -= g[p, q, n, r] * rdms.RDM2(p, q, t, r)
+                            A2e[idx1, idx2] -= g[p, q, n, r] * RDM2(
+                                p, q, t, r, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                            )
                         if t == n:
-                            A2e[idx1, idx2] -= g[p, q, r, m] * rdms.RDM2(p, q, r, u)
+                            A2e[idx1, idx2] -= g[p, q, r, m] * RDM2(
+                                p, q, r, u, num_inactive_orbs, num_active_orbs, rdm1, rdm2
+                            )
     return 1 / 2 * A1e + 1 / 4 * A2e
