@@ -4,81 +4,29 @@ import copy
 import re
 
 
-class a_op:
-    def __init__(self, spinless_idx: int, spin: str, dagger: bool) -> None:
-        """Initialize fermionic annihilation operator.
-
-        Args:
-            spinless_idx: Spatial orbital index.
-            spin: Alpha or beta spin.
-            dagger: If creation operator.
-        """
-        if spin not in ("alpha", "beta"):
-            raise ValueError(f'spin must be "alpha" or "beta" got {spin}')
-        self.spinless_idx = spinless_idx
-        self.idx = 2 * self.spinless_idx
-        self.dagger = dagger
-        self.spin = spin
-        if self.spin == "beta":
-            self.idx += 1
-
-
-def a_op_spin(spin_idx: int, dagger: bool) -> a_op:
-    """Get fermionic annihilation operator.
-
-    Args:
-        spin_idx: Spin orbital index.
-        dagger: If creation operator.
-
-    Returns:
-        Annihilation operator.
-    """
-    if spin_idx % 2 == 0:
-        return a_op(spin_idx // 2, "alpha", dagger)
-    return a_op(spin_idx // 2, "beta", dagger)
-
-
-def operator_string_to_key(operator_string: list[a_op]) -> str:
+def operator_to_qiskit_key(operator_string: tuple[tuple[int, bool], ...], remapping: dict[int, int]) -> str:
     """Make key string to index a fermionic operator in a dict structure.
 
     Args:
-        operator_string: Fermionic opreators.
-
-    Returns:
-        Dictionary key.
-    """
-    string_key = ""
-    for a in operator_string:
-        if a.dagger:
-            string_key += f"c{a.idx}"
-        else:
-            string_key += f"a{a.idx}"
-    return string_key
-
-
-def operator_to_qiskit_key(operator_string: list[a_op], remapping: dict[int, int]) -> str:
-    """Make key string to index a fermionic operator in a dict structure.
-
-    Args:
-        operator_string: Fermionic opreators.
+        operator_string: Fermionic operators.
         remapping: Map that takes indices from alpha,beta,alpha,beta
                    to alpha,alpha,beta,beta ordering.
 
     Returns:
         Dictionary key.
     """
-    string_key = ""
+    op_key = ""
     for a in operator_string:
-        if a.dagger:
-            string_key += f" +_{remapping[a.idx]}"
+        if a[1]:
+            op_key += f" +_{remapping[a[0]]}"
         else:
-            string_key += f" -_{remapping[a.idx]}"
-    return string_key[1:]
+            op_key += f" -_{remapping[a[0]]}"
+    return op_key[1:]
 
 
 def do_extended_normal_ordering(
     fermistring: FermionicOperator,
-) -> tuple[dict[str, list[a_op]], dict[str, float]]:
+) -> dict[tuple[tuple[int, bool], ...], float]:
     """Reorder fermionic operator string.
 
     The string will be ordered such that all creation operators are first,
@@ -92,10 +40,9 @@ def do_extended_normal_ordering(
     operator_queue = []
     factor_queue = []
     new_operators = {}
-    new_factors = {}
-    for key in fermistring.operators:
-        operator_queue.append(fermistring.operators[key])
-        factor_queue.append(fermistring.factors[key])
+    for key in fermistring.operators.keys():
+        operator_queue.append(list(key))
+        factor_queue.append(fermistring.operators[key])
     while len(operator_queue) > 0:
         next_operator = operator_queue.pop(0)
         factor = factor_queue.pop(0)
@@ -105,19 +52,21 @@ def do_extended_normal_ordering(
             changed = False
             is_zero = False
             while True:
+                if len(next_operator) == 0:
+                    break
                 a = next_operator[current_idx]
                 b = next_operator[current_idx + 1]
                 i = current_idx
                 j = current_idx + 1
-                if a.dagger and b.dagger:
-                    if a.idx == b.idx:
+                if a[1] and b[1]:
+                    if a[0] == b[0]:
                         is_zero = True
-                    elif a.idx < b.idx:
+                    elif a[0] < b[0]:
                         next_operator[i], next_operator[j] = next_operator[j], next_operator[i]
                         factor *= -1
                         changed = True
-                elif not a.dagger and b.dagger:
-                    if a.idx == b.idx:
+                elif not a[1] and b[1]:
+                    if a[0] == b[0]:
                         new_op = copy.copy(next_operator)
                         new_op.pop(j)
                         new_op.pop(i)
@@ -131,65 +80,50 @@ def do_extended_normal_ordering(
                         next_operator[i], next_operator[j] = next_operator[j], next_operator[i]
                         factor *= -1
                         changed = True
-                elif a.dagger and not b.dagger:
+                elif a[1] and not b[1]:
                     pass
-                else:
-                    if a.idx == b.idx:
-                        is_zero = True
-                    elif a.idx < b.idx:
-                        next_operator[i], next_operator[j] = next_operator[j], next_operator[i]
-                        factor *= -1
-                        changed = True
+                elif a[0] == b[0]:
+                    is_zero = True
+                elif a[0] < b[0]:
+                    next_operator[i], next_operator[j] = next_operator[j], next_operator[i]
+                    factor *= -1
+                    changed = True
                 current_idx += 1
                 if current_idx + 1 == len(next_operator) or is_zero:
                     break
             if not changed or is_zero:
                 if not is_zero:
-                    key_string = operator_string_to_key(next_operator)
-                    if key_string not in new_operators:
-                        new_operators[key_string] = next_operator
-                        new_factors[key_string] = factor
+                    op_key = tuple(next_operator)
+                    if op_key not in new_operators:
+                        new_operators[op_key] = factor
                     else:
-                        new_factors[key_string] += factor
-                        if abs(new_factors[key_string]) < 10**-14:
-                            del new_operators[key_string]
-                            del new_factors[key_string]
+                        new_operators[op_key] += factor
+                        if abs(new_operators[op_key]) < 10**-14:
+                            del new_operators[op_key]
                 break
-    return new_operators, new_factors
+    return new_operators
 
 
 class FermionicOperator:
+    __slots__ = ("operators",)
+
     def __init__(
-        self, annihilation_operator: dict[str, list[a_op]] | a_op, factor: dict[str, float] | float
+        self,
+        annihilation_operator: dict[tuple[tuple[int, bool], ...], float],
     ) -> None:
         """Initialize fermionic operator class.
 
-        Fermionic operators are defined via an annihilation_operator dictionary where each entry is one addend of the operator.
-        Each entry is a strings of annihilation operator specified via its string (key) and list of a_op (item).
-        The dictionary factor contains the factor for each of the addend of the fermionic operator.
+        Fermionic operators are defined via an annihilation_operator dictionary where each entry is one of the annihilation operators.
+        Each entry is a tuples (key) of an integer (spin orbital index) and a bool (dagger/not dagger)
+        which defines the keys and a float (value) that is the factor in front of the annihilation string.
 
         Args:
             annihilation_operator: Annihilation operator.
-            factor: Factor in front of operator.
         """
-        if isinstance(annihilation_operator, dict) and not isinstance(factor, dict):
-            raise ValueError(f"factor cannot be {type(factor)} when annihilation_operator is dict")
-        if not isinstance(annihilation_operator, dict) and isinstance(factor, float):
-            raise ValueError(f"factor cannot be dict when annihilation_operator is {type(a_op)}")
-        if not isinstance(annihilation_operator, dict) and not isinstance(factor, dict):
-            string_key = operator_string_to_key([annihilation_operator])
-            self.operators = {}
-            self.operators[string_key] = [annihilation_operator]
-            self.factors = {}
-            self.factors[string_key] = factor
-        elif isinstance(annihilation_operator, dict) and isinstance(factor, dict):
+        if isinstance(annihilation_operator, dict):
             self.operators = annihilation_operator
-            self.factors = factor
         else:
-            raise ValueError(
-                f"Could not assign operator of {type(annihilation_operator)} with factor of {type(factor)}"
-            )
-        self._operators_ab = None
+            raise ValueError(f"Could not assign operator of {type(annihilation_operator)}.")
 
     def __add__(self, fermistring: FermionicOperator) -> FermionicOperator:
         """Addition of two fermionic operators.
@@ -202,17 +136,32 @@ class FermionicOperator:
         """
         # Combine annihilation string entries of two FermionicOperators.
         operators = copy.copy(self.operators)
-        factors = copy.copy(self.factors)
-        for string_key in fermistring.operators.keys():
-            if string_key in operators.keys():
-                factors[string_key] += fermistring.factors[string_key]
-                if abs(factors[string_key]) < 10**-14:
-                    del factors[string_key]
-                    del operators[string_key]
+        for op_key in fermistring.operators.keys():
+            if op_key in operators.keys():
+                operators[op_key] += fermistring.operators[op_key]
+                if abs(operators[op_key]) < 10**-14:
+                    del operators[op_key]
             else:
-                operators[string_key] = fermistring.operators[string_key]
-                factors[string_key] = fermistring.factors[string_key]
-        return FermionicOperator(operators, factors)
+                operators[op_key] = fermistring.operators[op_key]
+        return FermionicOperator(operators)
+
+    def __iadd__(self, fermistring: FermionicOperator) -> FermionicOperator:
+        """Inplace addition of two fermionic operators.
+
+        Args:
+            fermistring: Fermionic operator.
+
+        Returns:
+            Updated fermionic operator.
+        """
+        for op_key in fermistring.operators.keys():
+            if op_key in self.operators.keys():
+                self.operators[op_key] += fermistring.operators[op_key]
+                if abs(self.operators[op_key]) < 10**-14:
+                    del self.operators[op_key]
+            else:
+                self.operators[op_key] = fermistring.operators[op_key]
+        return self
 
     def __sub__(self, fermistring: FermionicOperator) -> FermionicOperator:
         """Subtraction of two fermionic operators.
@@ -225,19 +174,35 @@ class FermionicOperator:
         """
         # Combine annihilation string entries of two FermionicOperators with relevant sign flip.
         operators = copy.copy(self.operators)
-        factors = copy.copy(self.factors)
-        for string_key in fermistring.operators.keys():
-            if string_key in operators.keys():
-                factors[string_key] -= fermistring.factors[string_key]
-                if abs(factors[string_key]) < 10**-14:
-                    del factors[string_key]
-                    del operators[string_key]
+        for op_key in fermistring.operators.keys():
+            if op_key in operators.keys():
+                operators[op_key] -= fermistring.operators[op_key]
+                if abs(operators[op_key]) < 10**-14:
+                    del operators[op_key]
             else:
-                operators[string_key] = fermistring.operators[string_key]
-                factors[string_key] = -fermistring.factors[string_key]
-        return FermionicOperator(operators, factors)
+                operators[op_key] = -fermistring.operators[op_key]
+        return FermionicOperator(operators)
 
-    def __mul__(self, fermistring: FermionicOperator) -> FermionicOperator:
+    def __isub__(self, fermistring: FermionicOperator) -> FermionicOperator:
+        """Inplace subtraction of two fermionic operators.
+
+        Args:
+            fermistring: Fermionic operator.
+
+        Returns:
+            Update fermionic operator.
+        """
+        # Combine annihilation string entries of two FermionicOperators with relevant sign flip.
+        for op_key in fermistring.operators.keys():
+            if op_key in self.operators.keys():
+                self.operators[op_key] -= fermistring.operators[op_key]
+                if abs(self.operators[op_key]) < 10**-14:
+                    del self.operators[op_key]
+            else:
+                self.operators[op_key] = -fermistring.operators[op_key]
+        return self
+
+    def __mul__(self, fermistring: FermionicOperator | float | int) -> FermionicOperator:
         """Multiplication of two fermionic operators.
 
         Args:
@@ -246,35 +211,68 @@ class FermionicOperator:
         Returns:
             New fermionic operator.
         """
-        operators: dict[str, list[a_op]] = {}
-        factors: dict[str, float] = {}
-        # Iterate over all strings in both FermionicOperators
-        for string_key1 in fermistring.operators.keys():
-            for string_key2 in self.operators.keys():
-                # Build new strings and factors via normal ordering of product of two strings
-                new_ops, new_facs = do_extended_normal_ordering(
-                    FermionicOperator(
-                        {
-                            string_key1
-                            + string_key2: self.operators[string_key2]
-                            + fermistring.operators[string_key1]
-                        },
-                        {
-                            string_key1
-                            + string_key2: self.factors[string_key2] * fermistring.factors[string_key1]
-                        },
+        if type(fermistring) in (float, int):
+            operators = copy.copy(self.operators)
+            for op_key in self.operators.keys():
+                # The name fermistring is misleading here.
+                operators[op_key] *= fermistring  # type: ignore
+        elif type(fermistring) is FermionicOperator:
+            operators = {}  # type: ignore
+            # Iterate over all strings in both FermionicOperators
+            for op_key1 in fermistring.operators.keys():
+                for op_key2 in self.operators.keys():
+                    # Build new strings and factors via normal ordering of product of two strings
+                    new_ops = do_extended_normal_ordering(
+                        FermionicOperator(
+                            {op_key2 + op_key1: self.operators[op_key2] * fermistring.operators[op_key1]}
+                        )
                     )
-                )
-                for str_key in new_ops:  # pylint: disable=C0206
-                    if str_key not in operators.keys():  # pylint: disable=C0201
-                        operators[str_key] = new_ops[str_key]
-                        factors[str_key] = new_facs[str_key]
-                    else:
-                        factors[str_key] += new_facs[str_key]
-                        if abs(factors[str_key]) < 10**-14:
-                            del factors[str_key]
-                            del operators[str_key]
-        return FermionicOperator(operators, factors)
+                    for op_key in new_ops.keys():
+                        if op_key not in operators.keys():
+                            operators[op_key] = new_ops[op_key]
+                        else:
+                            operators[op_key] += new_ops[op_key]
+                            if abs(operators[op_key]) < 10**-14:
+                                del operators[op_key]
+        else:
+            raise TypeError(f"Got unknown type of fermistring: {type(fermistring)}")
+        return FermionicOperator(operators)
+
+    def __imul__(self, fermistring: FermionicOperator | float | int) -> FermionicOperator:
+        """Inplace multiplication of two fermionic operators.
+
+        Args:
+            fermistring: Fermionic operator.
+
+        Returns:
+            Updated fermionic operator.
+        """
+        if type(fermistring) in (float, int):
+            for op_key in self.operators.keys():
+                # The name fermistring is misleading here.
+                self.operators[op_key] *= fermistring  # type: ignore
+        elif type(fermistring) is FermionicOperator:
+            operators: dict[tuple[tuple[int, bool], ...], float] = {}
+            # Iterate over all strings in both FermionicOperators
+            for op_key1 in fermistring.operators.keys():
+                for op_key2 in self.operators.keys():
+                    # Build new strings and factors via normal ordering of product of two strings
+                    new_ops = do_extended_normal_ordering(
+                        FermionicOperator(
+                            {op_key2 + op_key1: self.operators[op_key2] * fermistring.operators[op_key1]}
+                        )
+                    )
+                    for op_key in new_ops.keys():
+                        if op_key not in operators.keys():
+                            operators[op_key] = new_ops[op_key]
+                        else:
+                            operators[op_key] += new_ops[op_key]
+                            if abs(operators[op_key]) < 10**-14:
+                                del operators[op_key]
+            self.operators = operators
+        else:
+            raise TypeError(f"Got unknown type of fermistring: {type(fermistring)}")
+        return self
 
     def __rmul__(self, number: float) -> FermionicOperator:
         """Multiplication of number with fermionic operator.
@@ -286,11 +284,20 @@ class FermionicOperator:
             New fermionic operator.
         """
         operators = {}
-        factors = {}
-        for key_string in self.operators:
-            operators[key_string] = self.operators[key_string]
-            factors[key_string] = self.factors[key_string] * number
-        return FermionicOperator(operators, factors)
+        for op_key in self.operators.keys():
+            operators[op_key] = self.operators[op_key] * number
+        return FermionicOperator(operators)
+
+    def __neg__(self):
+        """Negate the factors in a fermionic operator.
+
+        Retunrs:
+            New fermionic operator.
+        """
+        operators = copy.copy(self.operators)
+        for op_key in self.operators.keys():
+            operators[op_key] = -operators[op_key]
+        return FermionicOperator(operators)
 
     @property
     def dagger(self) -> FermionicOperator:
@@ -300,18 +307,18 @@ class FermionicOperator:
             New fermionic operator.
         """
         operators = {}
-        factors = {}
-        for key_string in self.operators.keys():
+        for op_key in self.operators.keys():
             new_op = []
-            for op in reversed(self.operators[key_string]):
-                if op.dagger:
-                    new_op.append(a_op(op.spinless_idx, op.spin, False))
+            for op in reversed(op_key):
+                if op[1]:
+                    new_op.append((op[0], False))
                 else:
-                    new_op.append(a_op(op.spinless_idx, op.spin, True))
-            new_string_key = operator_string_to_key(new_op)
-            operators[new_string_key] = new_op
-            factors[new_string_key] = self.factors[key_string]
-        return FermionicOperator(operators, factors)
+                    new_op.append((op[0], True))
+            new_op_key = tuple(new_op)
+            operators[new_op_key] = self.operators[op_key]
+        # Do normal ordering of comlex conjugated operator.
+        operators_ordered = do_extended_normal_ordering(FermionicOperator(operators))
+        return FermionicOperator(operators_ordered)
 
     @property
     def operator_count(self) -> dict[int, int]:
@@ -321,13 +328,31 @@ class FermionicOperator:
             Number of operators of every length.
         """
         op_count = {}
-        for string_key in self.operators.keys():
-            op_lenght = len(self.operators[string_key])
+        for op_key in self.operators.keys():
+            op_lenght = len(op_key)
             if op_lenght not in op_count:
                 op_count[op_lenght] = 1
             else:
                 op_count[op_lenght] += 1
         return op_count
+
+    @property
+    def operators_readable(self) -> dict[str, float]:
+        """Get the operator in human readable format.
+
+        Returns:
+            Operator in humanreable format.
+        """
+        operator = {}
+        for string, fac in self.operators.items():
+            op_key = ""
+            for a in string:
+                if a[1]:
+                    op_key += f"c{a[0]}"
+                else:
+                    op_key += f"a{a[0]}"
+            operator[op_key] = fac
+        return operator
 
     def get_qiskit_form(self, num_orbs: int) -> dict[str, float]:
         """Get fermionic operator on qiskit form.
@@ -346,9 +371,9 @@ class FermionicOperator:
                 remapping[2 * i] = i
             else:
                 remapping[2 * i + 1 - 2 * num_orbs] = i
-        for key_string in self.operators:
-            qiskit_str = operator_to_qiskit_key(self.operators[key_string], remapping)
-            qiskit_form[qiskit_str] = self.factors[key_string]
+        for op_key in self.operators.keys():
+            qiskit_str = operator_to_qiskit_key(op_key, remapping)
+            qiskit_form[qiskit_str] = self.operators[op_key]
         return qiskit_form
 
     def get_folded_operator(
@@ -373,7 +398,7 @@ class FermionicOperator:
         Warning, multiplication of folded operators, might give wrong operators.
         (I have not quite figured out a good programming structure that will not allow multiplication after folding)
 
-        Note, that the indicies of the folded operator is remapped, such that idx=0 is the first index in the active space.
+        Note, that the indices of the folded operator is remapped, such that idx=0 is the first index in the active space.
 
         Args:
             num_inactive_orbs: Number of spatial inactive orbitals.
@@ -383,8 +408,7 @@ class FermionicOperator:
         Returns:
            Folded fermionic operator.
         """
-        operators = {}
-        factors: dict[str, float] = {}
+        operators: dict[tuple[tuple[int, bool], ...], float] = {}
         inactive_idx = []
         active_idx = []
         virtual_idx = []
@@ -398,7 +422,7 @@ class FermionicOperator:
                 virtual_idx.append(i)
 
         # Loop over string of annihilation operators
-        for key_string in self.operators.keys():
+        for op_key in self.operators.keys():
             virtual = []
             virtual_dagger = []
             inactive = []
@@ -407,23 +431,20 @@ class FermionicOperator:
             active_dagger = []
             fac = 1
             # Loop over individual annihilation operator and sort into spaces
-            for anni in self.operators[key_string]:
-                if anni.dagger:
-                    if anni.idx in inactive_idx:
-                        inactive_dagger.append(anni.idx)
-                    elif anni.idx in active_idx:
-                        active_dagger.append(
-                            a_op(anni.spinless_idx - num_inactive_orbs, anni.spin, anni.dagger)
-                        )
-                    elif anni.idx in virtual_idx:
-                        virtual_dagger.append(anni.idx)
-                else:
-                    if anni.idx in inactive_idx:
-                        inactive.append(anni.idx)
-                    elif anni.idx in active_idx:
-                        active.append(a_op(anni.spinless_idx - num_inactive_orbs, anni.spin, anni.dagger))
-                    elif anni.idx in virtual_idx:
-                        virtual.append(anni.idx)
+            for anni in op_key:
+                if anni[1]:
+                    if anni[0] in inactive_idx:
+                        inactive_dagger.append(anni[0])
+                    elif anni[0] in active_idx:
+                        active_dagger.append((anni[0] - 2 * num_inactive_orbs, anni[1]))
+                    elif anni[0] in virtual_idx:
+                        virtual_dagger.append(anni[0])
+                elif anni[0] in inactive_idx:
+                    inactive.append(anni[0])
+                elif anni[0] in active_idx:
+                    active.append((anni[0] - 2 * num_inactive_orbs, anni[1]))
+                elif anni[0] in virtual_idx:
+                    virtual.append(anni[0])
             # Any virtual indices will make the operator evaluate to zero.
             if len(virtual) != 0 or len(virtual_dagger) != 0:
                 continue
@@ -442,18 +463,18 @@ class FermionicOperator:
                 if i % 2 == 0:
                     ket_flip_fac *= -1
             fac *= ket_flip_fac
-            new_key = operator_string_to_key(active_op)
-            if new_key in factors:
-                factors[new_key] += fac * self.factors[key_string]
+            new_key = tuple(active_op)
+            if new_key in operators.keys():
+                operators[new_key] += fac * self.operators[op_key]
             else:
-                factors[new_key] = fac * self.factors[key_string]
-                operators[new_key] = active_op
-        return FermionicOperator(operators, factors)
+                operators[new_key] = fac * self.operators[op_key]
+        return FermionicOperator(operators)
 
     def get_info(self) -> tuple[list[list[int]], list[list[int]], list[float]]:
         """Return operator excitation in ordered strings with coefficient."""
-        excitations = list(self.factors.keys())
-        coefficients = list(self.factors.values())
+        operator = self.operators_readable
+        excitations = list(operator.keys())
+        coefficients = list(operator.values())
         creation = []
         annihilation = []
         for op_string in excitations:
@@ -465,63 +486,3 @@ class FermionicOperator:
             creation.append(c)
             annihilation.append(a)
         return annihilation, creation, coefficients
-
-
-def get_determinant_expansion_from_operator_on_HF(
-    operator: FermionicOperator,
-    num_active_orbs: int,
-    num_active_elec_alpha: int,
-    num_active_elec_beta: int,
-) -> tuple[list[float], list[str]]:
-    """Erik could have described the function.
-
-    Args:
-        operator: Fermionic operator.
-        num_active_orbs: Number of spatial active orbitals.
-        num_active_elec_alpha: Number of active alpha electrons.
-        num_active_elec_beta: Number of active beta electrons.
-    """
-    hf_det_ = ""
-    for i in range(2 * num_active_orbs):
-        if i % 2 == 0 and i // 2 < num_active_elec_alpha:
-            hf_det_ += "1"
-            continue
-        if i % 2 == 1 and i // 2 < num_active_elec_beta:
-            hf_det_ += "1"
-            continue
-        hf_det_ += "0"
-    hf_det = int(hf_det_, 2)
-
-    coeffs: list[float] = []
-    dets: list[str] = []
-    parity_check = {0: 0}
-    num = 0
-    for i in range(2 * num_active_orbs - 1, -1, -1):
-        num += 2**i
-        parity_check[2 * num_active_orbs - i] = num
-    for fermi_label in operator.factors:
-        det = hf_det
-        phase_changes = 0
-        for fermi_op in operator.operators[fermi_label][::-1]:
-            orb_idx = fermi_op.idx
-            nth_bit = (det >> 2 * num_active_orbs - 1 - orb_idx) & 1
-            if nth_bit == 0 and fermi_op.dagger:
-                det = det ^ 2 ** (2 * num_active_orbs - 1 - orb_idx)
-                phase_changes += (det & parity_check[orb_idx]).bit_count()
-            elif nth_bit == 1 and fermi_op.dagger:
-                break
-            elif nth_bit == 0 and not fermi_op.dagger:
-                break
-            elif nth_bit == 1 and not fermi_op.dagger:
-                det = det ^ 2 ** (2 * num_active_orbs - 1 - orb_idx)
-                phase_changes += (det & parity_check[orb_idx]).bit_count()
-        else:  # nobreak
-            val = operator.factors[fermi_label] * (-1) ** phase_changes
-            det_str = format(det, f"0{2 * num_active_orbs}b")
-            if det_str in dets:
-                idx = dets.index(det_str)
-                coeffs[idx] += val
-            else:
-                dets.append(format(det, f"0{2 * num_active_orbs}b"))
-                coeffs.append(val)
-    return coeffs, dets
