@@ -6,14 +6,13 @@ from slowquant.molecularintegrals.integralfunctions import (
     one_electron_integral_transform,
 )
 from slowquant.unitary_coupled_cluster.density_matrix import (
-    ReducedDenstiyMatrix,
     get_orbital_gradient_response,
     get_orbital_response_property_gradient,
 )
 from slowquant.unitary_coupled_cluster.linear_response.lr_baseclass import (
     LinearResponseBaseClass,
 )
-from slowquant.unitary_coupled_cluster.operator_matrix import (
+from slowquant.unitary_coupled_cluster.operator_state_algebra import (
     expectation_value,
     propagate_state,
 )
@@ -47,25 +46,19 @@ class LinearResponseUCC(LinearResponseBaseClass):
             self.wf.num_virtual_orbs,
         )
 
-        rdms = ReducedDenstiyMatrix(
-            self.wf.num_inactive_orbs,
-            self.wf.num_active_orbs,
-            self.wf.num_virtual_orbs,
-            self.wf.rdm1,
-            rdm2=self.wf.rdm2,
-        )
         idx_shift = len(self.q_ops)
         print("Gs", len(self.G_ops))
         print("qs", len(self.q_ops))
-        grad = get_orbital_gradient_response(  # proj-q and naive-q lead to same working equations
-            rdms,
-            self.wf.h_mo,
-            self.wf.g_mo,
-            self.wf.kappa_no_activeactive_idx,
-            self.wf.num_inactive_orbs,
-            self.wf.num_active_orbs,
-        )
-        if len(grad) != 0:
+        if len(self.q_ops) != 0:
+            grad = get_orbital_gradient_response(  # proj-q and naive-q lead to same working equations
+                self.wf.h_mo,
+                self.wf.g_mo,
+                self.wf.kappa_no_activeactive_idx,
+                self.wf.num_inactive_orbs,
+                self.wf.num_active_orbs,
+                self.wf.rdm1,
+                self.wf.rdm2,
+            )
             print("idx, max(abs(grad orb)):", np.argmax(np.abs(grad)), np.max(np.abs(grad)))
             if np.max(np.abs(grad)) > 10**-3:
                 raise ValueError("Large Gradient detected in q of ", np.max(np.abs(grad)))
@@ -159,28 +152,6 @@ class LinearResponseUCC(LinearResponseBaseClass):
                     HGJ_ket,
                     *self.index_info,
                 )
-                # - <0| GId GJ |0> * E
-                val -= (
-                    expectation_value(
-                        GI_ket,
-                        [],
-                        GJ_ket,
-                        *self.index_info,
-                    )
-                    * self.wf.energy_elec
-                )
-                # - <0| GId |0> * <0| H GJ |0>
-                val -= expectation_value(
-                    GI_ket,
-                    [],
-                    self.wf.ci_coeffs,
-                    *self.index_info,
-                ) * expectation_value(
-                    self.wf.ci_coeffs,
-                    [],
-                    HGJ_ket,
-                    *self.index_info,
-                )
                 # <0 | GId |0> * <0| GJ |0> * E
                 val += (
                     expectation_value(
@@ -197,19 +168,85 @@ class LinearResponseUCC(LinearResponseBaseClass):
                     )
                     * self.wf.energy_elec
                 )
+                # - <0| GId GJ |0> * E
+                val -= (
+                    expectation_value(
+                        GI_ket,
+                        [],
+                        GJ_ket,
+                        *self.index_info,
+                    )
+                    * self.wf.energy_elec
+                )
+                # - 1/2*<0| GId |0> * <0| H GJ |0>
+                val -= (
+                    1
+                    / 2
+                    * expectation_value(
+                        GI_ket,
+                        [],
+                        self.wf.ci_coeffs,
+                        *self.index_info,
+                    )
+                    * expectation_value(
+                        self.wf.ci_coeffs,
+                        [],
+                        HGJ_ket,
+                        *self.index_info,
+                    )
+                )
+                # - 1/2*<0| GJ |0> * <0| GId H |0>
+                val -= (
+                    1
+                    / 2
+                    * expectation_value(
+                        self.wf.ci_coeffs,
+                        [],
+                        GJ_ket,
+                        *self.index_info,
+                    )
+                    * expectation_value(
+                        GI_ket,
+                        [self.H_1i_1a],
+                        self.wf.ci_coeffs,
+                        *self.index_info,
+                    )
+                )
                 self.A[i + idx_shift, j + idx_shift] = self.A[j + idx_shift, i + idx_shift] = val
                 # Make B
-                # <0| GId H |0> * <0| GJd |0>
-                val = expectation_value(
-                    self.wf.ci_coeffs,
-                    [GI.dagger, self.H_0i_0a],
-                    self.wf.ci_coeffs,
-                    *self.index_info,
-                ) * expectation_value(
-                    self.wf.ci_coeffs,
-                    [GJ.dagger],
-                    self.wf.ci_coeffs,
-                    *self.index_info,
+                # 1/2<0| GId H |0> * <0| GJd |0>
+                val = (
+                    1
+                    / 2
+                    * expectation_value(
+                        self.wf.ci_coeffs,
+                        [GI.dagger, self.H_0i_0a],
+                        self.wf.ci_coeffs,
+                        *self.index_info,
+                    )
+                    * expectation_value(
+                        self.wf.ci_coeffs,
+                        [GJ.dagger],
+                        self.wf.ci_coeffs,
+                        *self.index_info,
+                    )
+                )
+                # 1/2<0| GJd H |0> * <0| GId |0>
+                val += (
+                    1
+                    / 2
+                    * expectation_value(
+                        self.wf.ci_coeffs,
+                        [GJ.dagger, self.H_0i_0a],
+                        self.wf.ci_coeffs,
+                        *self.index_info,
+                    )
+                    * expectation_value(
+                        self.wf.ci_coeffs,
+                        [GI.dagger],
+                        self.wf.ci_coeffs,
+                        *self.index_info,
+                    )
                 )
                 # - <0| GId |0> * <0| GJd |0> * E
                 val -= (
@@ -262,16 +299,9 @@ class LinearResponseUCC(LinearResponseBaseClass):
         if len(dipole_integrals) != 3:
             raise ValueError(f"Expected 3 dipole integrals got {len(dipole_integrals)}")
         number_excitations = len(self.excitation_energies)
-        rdms = ReducedDenstiyMatrix(
-            self.wf.num_inactive_orbs,
-            self.wf.num_active_orbs,
-            self.wf.num_virtual_orbs,
-            self.wf.rdm1,
-            rdm2=self.wf.rdm2,
-        )
-        mux = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[0])
-        muy = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[1])
-        muz = one_electron_integral_transform(self.wf.c_trans, dipole_integrals[2])
+        mux = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[0])
+        muy = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[1])
+        muz = one_electron_integral_transform(self.wf.c_mo, dipole_integrals[2])
         mux_op = one_elec_op_0i_0a(
             mux,
             self.wf.num_inactive_orbs,
@@ -295,36 +325,40 @@ class LinearResponseUCC(LinearResponseBaseClass):
         muzd_ket = propagate_state([muz_op.dagger], self.wf.ci_coeffs, *self.index_info)
         transition_dipoles = np.zeros((number_excitations, 3))
         for state_number in range(number_excitations):
-            q_part_x = get_orbital_response_property_gradient(
-                rdms,
-                mux,
-                self.wf.kappa_no_activeactive_idx,
-                self.wf.num_inactive_orbs,
-                self.wf.num_active_orbs,
-                self.normed_response_vectors,
-                state_number,
-                number_excitations,
-            )
-            q_part_y = get_orbital_response_property_gradient(
-                rdms,
-                muy,
-                self.wf.kappa_no_activeactive_idx,
-                self.wf.num_inactive_orbs,
-                self.wf.num_active_orbs,
-                self.normed_response_vectors,
-                state_number,
-                number_excitations,
-            )
-            q_part_z = get_orbital_response_property_gradient(
-                rdms,
-                muz,
-                self.wf.kappa_no_activeactive_idx,
-                self.wf.num_inactive_orbs,
-                self.wf.num_active_orbs,
-                self.normed_response_vectors,
-                state_number,
-                number_excitations,
-            )
+            q_part_x = 0.0
+            q_part_y = 0.0
+            q_part_z = 0.0
+            if len(self.q_ops) != 0:
+                q_part_x = get_orbital_response_property_gradient(
+                    mux,
+                    self.wf.kappa_no_activeactive_idx,
+                    self.wf.num_inactive_orbs,
+                    self.wf.num_active_orbs,
+                    self.wf.rdm1,
+                    self.normed_response_vectors,
+                    state_number,
+                    number_excitations,
+                )
+                q_part_y = get_orbital_response_property_gradient(
+                    muy,
+                    self.wf.kappa_no_activeactive_idx,
+                    self.wf.num_inactive_orbs,
+                    self.wf.num_active_orbs,
+                    self.wf.rdm1,
+                    self.normed_response_vectors,
+                    state_number,
+                    number_excitations,
+                )
+                q_part_z = get_orbital_response_property_gradient(
+                    muz,
+                    self.wf.kappa_no_activeactive_idx,
+                    self.wf.num_inactive_orbs,
+                    self.wf.num_active_orbs,
+                    self.wf.rdm1,
+                    self.normed_response_vectors,
+                    state_number,
+                    number_excitations,
+                )
             g_part_x = 0.0
             g_part_y = 0.0
             g_part_z = 0.0
