@@ -2,22 +2,39 @@ import numpy as np
 
 from slowquant.unitary_coupled_cluster.fermionic_operator import (
     FermionicOperator,
-    a_op,
-    a_op_spin,
 )
 
 
-def anni_spin(p: int, dagger: bool) -> FermionicOperator:
+def a_op(spinless_idx: int, spin: str, dagger: bool) -> FermionicOperator:
     """Construct annihilation/creation operator.
 
     Args:
-        p: Spin orbital index.
+        spinless_idx: Spatial orbital index.
+        spin: Alpha or beta spin.
         dagger: If creation operator.
 
     Returns:
         Annihilation/creation operator.
     """
-    return FermionicOperator(a_op_spin(p, dagger=dagger), 1)
+    if spin not in ("alpha", "beta"):
+        raise ValueError(f'spin must be "alpha" or "beta" got {spin}')
+    idx = 2 * spinless_idx
+    if spin == "beta":
+        idx += 1
+    return FermionicOperator({((idx, dagger),): 1})
+
+
+def a_op_spin(spin_idx: int, dagger: bool) -> FermionicOperator:
+    """Construct annihilation/creation operator.
+
+    Args:
+        spin_idx: Spin orbital index.
+        dagger: If creation operator.
+
+    Returns:
+        Annihilation/creation operator.
+    """
+    return FermionicOperator({((spin_idx, dagger),): 1})
 
 
 def Epq(p: int, q: int) -> FermionicOperator:
@@ -33,12 +50,11 @@ def Epq(p: int, q: int) -> FermionicOperator:
     Returns:
         Singlet one-electron excitation operator.
     """
-    E = FermionicOperator(a_op(p, "alpha", dagger=True), 1) * FermionicOperator(
-        a_op(q, "alpha", dagger=False), 1
-    )
-    E += FermionicOperator(a_op(p, "beta", dagger=True), 1) * FermionicOperator(
-        a_op(q, "beta", dagger=False), 1
-    )
+    E = a_op(p, "alpha", dagger=True)
+    E *= a_op(q, "alpha", dagger=False)
+    tmp = a_op(p, "beta", dagger=True)
+    tmp *= a_op(q, "beta", dagger=False)
+    E += tmp
     return E
 
 
@@ -57,9 +73,11 @@ def epqrs(p: int, q: int, r: int, s: int) -> FermionicOperator:
     Returns:
         Singlet two-electron excitation operator.
     """
+    op = Epq(p, q)
+    op *= Epq(r, s)
     if q == r:
-        return Epq(p, q) * Epq(r, s) - Epq(p, s)
-    return Epq(p, q) * Epq(r, s)
+        op -= Epq(p, s)
+    return op
 
 
 def Eminuspq(p: int, q: int) -> FermionicOperator:
@@ -75,7 +93,9 @@ def Eminuspq(p: int, q: int) -> FermionicOperator:
     Returns:
         Singlet one-electron excitation operator.
     """
-    return Epq(p, q) - Epq(q, p)
+    op = Epq(p, q)
+    op -= Epq(q, p)
+    return op
 
 
 def commutator(A: FermionicOperator, B: FermionicOperator) -> FermionicOperator:
@@ -94,24 +114,35 @@ def commutator(A: FermionicOperator, B: FermionicOperator) -> FermionicOperator:
     return A * B - B * A
 
 
-def double_commutator(A: FermionicOperator, B: FermionicOperator, C: FermionicOperator) -> FermionicOperator:
+def double_commutator(
+    A: FermionicOperator, B: FermionicOperator, C: FermionicOperator, do_symmetrized: bool = False
+) -> FermionicOperator:
     r"""Construct operator double commutator.
 
     .. math::
         \left[\hat{A},\left[\hat{B},\hat{C}\right]\right] = \hat{A}\hat{B}\hat{C} - \hat{A}\hat{C}\hat{B} - \hat{B}\hat{C}\hat{A} + \hat{C}\hat{B}\hat{A}
 
+    or for the symmetrized version,
+
+    .. math::
+        \left[\hat{A},\hat{B},\hat{C}\right] =
+        \hat{A}\hat{H}\hat{B} + \hat{B}\hat{H}\hat{A} - \frac{1}{2}\left(\hat{A}\hat{B}\hat{H} + \hat{H}\hat{B}\hat{A} + \hat{B}\hat{A}\hat{H} + \hat{H}\hat{A}\hat{B}\right)
+
     Args:
         A: Fermionic operator.
         B: Fermionic operator.
         C: Fermionic operator.
+        do_symmetrized: Do symmetrized double commutator (default: False).
 
     Returns:
         Operator from double commutator.
     """
+    if do_symmetrized:
+        return A * B * C + C * B * A - 1 / 2 * (A * C * B + B * C * A + C * A * B + B * A * C)
     return A * B * C - A * C * B - B * C * A + C * B * A
 
 
-def G1(i: int, a: int) -> FermionicOperator:
+def G1(i: int, a: int, return_anti_hermitian: bool = False) -> FermionicOperator:
     r"""Construct one-electron excitation operator.
 
     .. math::
@@ -120,14 +151,19 @@ def G1(i: int, a: int) -> FermionicOperator:
     Args:
         i: Spin orbital index.
         a: Spin orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger
 
     Returns:
-        One-elecetron excitation operator.
+        One-electron excitation operator.
     """
-    return FermionicOperator(a_op_spin(a, dagger=True), 1) * FermionicOperator(a_op_spin(i, dagger=False), 1)
+    op = a_op_spin(a, dagger=True)
+    op *= a_op_spin(i, dagger=False)
+    if return_anti_hermitian:
+        op -= op.dagger
+    return op
 
 
-def G2(i: int, j: int, a: int, b: int) -> FermionicOperator:
+def G2(i: int, j: int, a: int, b: int, return_anti_hermitian: bool = False) -> FermionicOperator:
     r"""Construct two-electron excitation operator.
 
     .. math::
@@ -138,19 +174,23 @@ def G2(i: int, j: int, a: int, b: int) -> FermionicOperator:
         j: Spin orbital index.
         a: Spin orbital index.
         b: Spin orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        Two-elecetron excitation operator.
+        Two-electron excitation operator.
     """
-    return (
-        FermionicOperator(a_op_spin(a, dagger=True), 1)
-        * FermionicOperator(a_op_spin(b, dagger=True), 1)
-        * FermionicOperator(a_op_spin(j, dagger=False), 1)
-        * FermionicOperator(a_op_spin(i, dagger=False), 1)
-    )
+    op = a_op_spin(a, dagger=True)
+    op *= a_op_spin(b, dagger=True)
+    op *= a_op_spin(j, dagger=False)
+    op *= a_op_spin(i, dagger=False)
+    if return_anti_hermitian:
+        op -= op.dagger
+    return op
 
 
-def G3(i: int, j: int, k: int, a: int, b: int, c: int) -> FermionicOperator:
+def G3(
+    i: int, j: int, k: int, a: int, b: int, c: int, return_anti_hermitian: bool = False
+) -> FermionicOperator:
     r"""Construct three-electron excitation operator.
 
     .. math::
@@ -163,21 +203,25 @@ def G3(i: int, j: int, k: int, a: int, b: int, c: int) -> FermionicOperator:
         a: Spin orbital index.
         b: Spin orbital index.
         c: Spin orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        Three-elecetron excitation operator.
+        Three-electron excitation operator.
     """
-    return (
-        FermionicOperator(a_op_spin(a, dagger=True), 1)
-        * FermionicOperator(a_op_spin(b, dagger=True), 1)
-        * FermionicOperator(a_op_spin(c, dagger=True), 1)
-        * FermionicOperator(a_op_spin(k, dagger=False), 1)
-        * FermionicOperator(a_op_spin(j, dagger=False), 1)
-        * FermionicOperator(a_op_spin(i, dagger=False), 1)
-    )
+    op = a_op_spin(a, dagger=True)
+    op *= a_op_spin(b, dagger=True)
+    op *= a_op_spin(c, dagger=True)
+    op *= a_op_spin(k, dagger=False)
+    op *= a_op_spin(j, dagger=False)
+    op *= a_op_spin(i, dagger=False)
+    if return_anti_hermitian:
+        op -= op.dagger
+    return op
 
 
-def G4(i: int, j: int, k: int, l: int, a: int, b: int, c: int, d: int) -> FermionicOperator:
+def G4(
+    i: int, j: int, k: int, l: int, a: int, b: int, c: int, d: int, return_anti_hermitian: bool = False
+) -> FermionicOperator:
     r"""Construct four-electron excitation operator.
 
     .. math::
@@ -192,23 +236,37 @@ def G4(i: int, j: int, k: int, l: int, a: int, b: int, c: int, d: int) -> Fermio
         b: Spin orbital index.
         c: Spin orbital index.
         d: Spin orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        Four-elecetron excitation operator.
+        Four-electron excitation operator.
     """
-    return (
-        FermionicOperator(a_op_spin(a, dagger=True), 1)
-        * FermionicOperator(a_op_spin(b, dagger=True), 1)
-        * FermionicOperator(a_op_spin(c, dagger=True), 1)
-        * FermionicOperator(a_op_spin(d, dagger=True), 1)
-        * FermionicOperator(a_op_spin(l, dagger=False), 1)
-        * FermionicOperator(a_op_spin(k, dagger=False), 1)
-        * FermionicOperator(a_op_spin(j, dagger=False), 1)
-        * FermionicOperator(a_op_spin(i, dagger=False), 1)
-    )
+    op = a_op_spin(a, dagger=True)
+    op *= a_op_spin(b, dagger=True)
+    op *= a_op_spin(c, dagger=True)
+    op *= a_op_spin(d, dagger=True)
+    op *= a_op_spin(l, dagger=False)
+    op *= a_op_spin(k, dagger=False)
+    op *= a_op_spin(j, dagger=False)
+    op *= a_op_spin(i, dagger=False)
+    if return_anti_hermitian:
+        op -= op.dagger
+    return op
 
 
-def G5(i: int, j: int, k: int, l: int, m: int, a: int, b: int, c: int, d: int, e: int) -> FermionicOperator:
+def G5(
+    i: int,
+    j: int,
+    k: int,
+    l: int,
+    m: int,
+    a: int,
+    b: int,
+    c: int,
+    d: int,
+    e: int,
+    return_anti_hermitian: bool = False,
+) -> FermionicOperator:
     r"""Construct five-electron excitation operator.
 
     .. math::
@@ -225,26 +283,40 @@ def G5(i: int, j: int, k: int, l: int, m: int, a: int, b: int, c: int, d: int, e
         c: Spin orbital index.
         d: Spin orbital index.
         e: Spin orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        Five-elecetron excitation operator.
+        Five-electron excitation operator.
     """
-    return (
-        FermionicOperator(a_op_spin(a, dagger=True), 1)
-        * FermionicOperator(a_op_spin(b, dagger=True), 1)
-        * FermionicOperator(a_op_spin(c, dagger=True), 1)
-        * FermionicOperator(a_op_spin(d, dagger=True), 1)
-        * FermionicOperator(a_op_spin(e, dagger=True), 1)
-        * FermionicOperator(a_op_spin(m, dagger=False), 1)
-        * FermionicOperator(a_op_spin(l, dagger=False), 1)
-        * FermionicOperator(a_op_spin(k, dagger=False), 1)
-        * FermionicOperator(a_op_spin(j, dagger=False), 1)
-        * FermionicOperator(a_op_spin(i, dagger=False), 1)
-    )
+    op = a_op_spin(a, dagger=True)
+    op *= a_op_spin(b, dagger=True)
+    op *= a_op_spin(c, dagger=True)
+    op *= a_op_spin(d, dagger=True)
+    op *= a_op_spin(e, dagger=True)
+    op *= a_op_spin(m, dagger=False)
+    op *= a_op_spin(l, dagger=False)
+    op *= a_op_spin(k, dagger=False)
+    op *= a_op_spin(j, dagger=False)
+    op *= a_op_spin(i, dagger=False)
+    if return_anti_hermitian:
+        op -= op.dagger
+    return op
 
 
 def G6(
-    i: int, j: int, k: int, l: int, m: int, n: int, a: int, b: int, c: int, d: int, e: int, f: int
+    i: int,
+    j: int,
+    k: int,
+    l: int,
+    m: int,
+    n: int,
+    a: int,
+    b: int,
+    c: int,
+    d: int,
+    e: int,
+    f: int,
+    return_anti_hermitian: bool = False,
 ) -> FermionicOperator:
     r"""Construct six-electron excitation operator.
 
@@ -265,27 +337,29 @@ def G6(
         d: Spin orbital index.
         e: Spin orbital index.
         f: Spin orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        Six-elecetron excitation operator.
+        Six-electron excitation operator.
     """
-    return (
-        FermionicOperator(a_op_spin(a, dagger=True), 1)
-        * FermionicOperator(a_op_spin(b, dagger=True), 1)
-        * FermionicOperator(a_op_spin(c, dagger=True), 1)
-        * FermionicOperator(a_op_spin(d, dagger=True), 1)
-        * FermionicOperator(a_op_spin(e, dagger=True), 1)
-        * FermionicOperator(a_op_spin(f, dagger=True), 1)
-        * FermionicOperator(a_op_spin(n, dagger=False), 1)
-        * FermionicOperator(a_op_spin(m, dagger=False), 1)
-        * FermionicOperator(a_op_spin(l, dagger=False), 1)
-        * FermionicOperator(a_op_spin(k, dagger=False), 1)
-        * FermionicOperator(a_op_spin(j, dagger=False), 1)
-        * FermionicOperator(a_op_spin(i, dagger=False), 1)
-    )
+    op = a_op_spin(a, dagger=True)
+    op *= a_op_spin(b, dagger=True)
+    op *= a_op_spin(c, dagger=True)
+    op *= a_op_spin(d, dagger=True)
+    op *= a_op_spin(e, dagger=True)
+    op *= a_op_spin(f, dagger=True)
+    op *= a_op_spin(n, dagger=False)
+    op *= a_op_spin(m, dagger=False)
+    op *= a_op_spin(l, dagger=False)
+    op *= a_op_spin(k, dagger=False)
+    op *= a_op_spin(j, dagger=False)
+    op *= a_op_spin(i, dagger=False)
+    if return_anti_hermitian:
+        op -= op.dagger
+    return op
 
 
-def G1_sa(i: int, a: int) -> FermionicOperator:
+def G1_sa(i: int, a: int, return_anti_hermitian: bool = False) -> FermionicOperator:
     r"""Construct singlet one-electron spin-adapted excitation operator.
 
     .. math::
@@ -294,14 +368,18 @@ def G1_sa(i: int, a: int) -> FermionicOperator:
     Args:
         i: Spatial orbital index.
         a: Spatial orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        Singlet one-elecetron spin-adapted excitation operator.
+        Singlet one-electron spin-adapted excitation operator.
     """
-    return 2 ** (-1 / 2) * Epq(a, i)
+    op = 2 ** (-1 / 2) * Epq(a, i)
+    if return_anti_hermitian:
+        op -= op.dagger
+    return op
 
 
-def G2_1_sa(i: int, j: int, a: int, b: int) -> FermionicOperator:
+def G2_1_sa(i: int, j: int, a: int, b: int, return_anti_hermitian: bool = False) -> FermionicOperator:
     r"""Construct first singlet two-electron spin-adapted excitation operator.
 
     .. math::
@@ -312,19 +390,23 @@ def G2_1_sa(i: int, j: int, a: int, b: int) -> FermionicOperator:
         j: Spatial orbital index.
         a: Spatial orbital index.
         b: Spatial orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        First singlet two-elecetron spin-adapted excitation operator.
+        First singlet two-electron spin-adapted excitation operator.
     """
     fac = 1
     if a == b:
         fac *= 2
     if i == j:
         fac *= 2
-    return 1 / 2 * (fac) ** (-1 / 2) * (Epq(a, i) * Epq(b, j) + Epq(a, j) * Epq(b, i))
+    op = 1 / 2 * (fac) ** (-1 / 2) * (Epq(a, i) * Epq(b, j) + Epq(a, j) * Epq(b, i))
+    if return_anti_hermitian:
+        op -= op.dagger
+    return op
 
 
-def G2_2_sa(i: int, j: int, a: int, b: int) -> FermionicOperator:
+def G2_2_sa(i: int, j: int, a: int, b: int, return_anti_hermitian: bool = False) -> FermionicOperator:
     r"""Construct second singlet two-electron spin-adapted excitation operator.
 
     .. math::
@@ -335,11 +417,15 @@ def G2_2_sa(i: int, j: int, a: int, b: int) -> FermionicOperator:
         j: Spatial orbital index.
         a: Spatial orbital index.
         b: Spatial orbital index.
+        return_anti_hermitian: Return anti-hermitian version of operator, i.e. op - op^\dagger.
 
     Returns:
-        Second singlet two-elecetron spin-adapted excitation operator.
+        Second singlet two-electron spin-adapted excitation operator.
     """
-    return 1 / (2 * 3 ** (1 / 2)) * (Epq(a, i) * Epq(b, j) - Epq(a, j) * Epq(b, i))
+    op = 1 / (2 * 3 ** (1 / 2)) * (Epq(a, i) * Epq(b, j) - Epq(a, j) * Epq(b, i))
+    if return_anti_hermitian:
+        op -= op.dagger
+    return op
 
 
 def hamiltonian_full_space(h_mo: np.ndarray, g_mo: np.ndarray, num_orbs: int) -> FermionicOperator:
@@ -356,7 +442,7 @@ def hamiltonian_full_space(h_mo: np.ndarray, g_mo: np.ndarray, num_orbs: int) ->
     Returns:
         Hamiltonian operator in full-space.
     """
-    H_operator = FermionicOperator({}, {})
+    H_operator = FermionicOperator({})
     for p in range(num_orbs):
         for q in range(num_orbs):
             if abs(h_mo[p, q]) < 10**-14:
@@ -387,9 +473,9 @@ def hamiltonian_0i_0a(
         num_active_orbs: Number of active orbitals in spatial basis.
 
     Returns:
-        Energy Hamilonian fermionic operator.
+        Energy Hamiltonian fermionic operator.
     """
-    hamiltonian_operator = FermionicOperator({}, {})
+    hamiltonian_operator = FermionicOperator({})
     # Inactive one-electron
     for i in range(num_inactive_orbs):
         if abs(h_mo[i, i]) > 10**-14:
@@ -445,10 +531,10 @@ def hamiltonian_1i_1a(
         num_virtual_orbs: Number of virtual orbitals in spatial basis.
 
     Returns:
-        Modified Hamilonian fermionic operator.
+        Modified Hamiltonian fermionic operator.
     """
     num_orbs = num_inactive_orbs + num_active_orbs + num_virtual_orbs
-    hamiltonian_operator = FermionicOperator({}, {})
+    hamiltonian_operator = FermionicOperator({})
     virtual_start = num_inactive_orbs + num_active_orbs
     for p in range(num_orbs):
         for q in range(num_orbs):
@@ -514,10 +600,10 @@ def hamiltonian_2i_2a(
         num_virtual_orbs: Number of virtual orbitals in spatial basis.
 
     Returns:
-        Modified Hamilonian fermionic operator.
+        Modified Hamiltonian fermionic operator.
     """
     num_orbs = num_inactive_orbs + num_active_orbs + num_virtual_orbs
-    hamiltonian_operator = FermionicOperator({}, {})
+    hamiltonian_operator = FermionicOperator({})
     virtual_start = num_inactive_orbs + num_active_orbs
     for p in range(num_orbs):
         for q in range(num_orbs):
@@ -575,7 +661,7 @@ def one_elec_op_full_space(ints_mo: np.ndarray, num_orbs: int) -> FermionicOpera
     Returns:
         One-electron operator in full-space.
     """
-    one_elec_op = FermionicOperator({}, {})
+    one_elec_op = FermionicOperator({})
     for p in range(num_orbs):
         for q in range(num_orbs):
             if abs(ints_mo[p, q]) > 10**-14:
@@ -594,7 +680,7 @@ def one_elec_op_0i_0a(ints_mo: np.ndarray, num_inactive_orbs: int, num_active_or
     Returns:
         One-electron operator for active-space.
     """
-    one_elec_op = FermionicOperator({}, {})
+    one_elec_op = FermionicOperator({})
     # Inactive one-electron
     for i in range(num_inactive_orbs):
         if abs(ints_mo[i, i]) > 10**-14:
@@ -622,7 +708,7 @@ def one_elec_op_1i_1a(
         Modified one-electron operator.
     """
     num_orbs = num_inactive_orbs + num_active_orbs + num_virtual_orbs
-    one_elec_op = FermionicOperator({}, {})
+    one_elec_op = FermionicOperator({})
     virtual_start = num_inactive_orbs + num_active_orbs
     for p in range(num_orbs):
         for q in range(num_orbs):
